+++ docs/docusaurus.config.ts
@@ -34,7 +34,6 @@ const config: Config = {
       {
         docs: {
           sidebarPath: "./sidebars.ts",
-          editUrl: "https://github.com/tejasbenibagde/fuderu/tree/main/docs/",
         },
         blog: false,
         theme: {